fixed memory error (parent not set to null during delete), got rid of a lot of compiler warnings (unqualified calls to move and implicit casts to int), fixed a bug in scrollpane where beginBaseAnimation() wouldn't work during a drag operation

+++ src/CppTweener.cpp
@@ -353,7 +353,7 @@ namespace tween {
 		total_tweens = tweens.size();
 		float t = 0 ;
 		float d = 0;
-		int  dif = (currentMillis - lastTime);
+		long  dif = (currentMillis - lastTime);
 		
 		for (tweensIT = tweens.begin();  tweensIT != tweens.end(); ++tweensIT ) {
